se movieron las excepciones a un archivo aparte
